core * Some overlapping packets reintroduced * a little work
git-svn-id: http://svn.l2jfree.com/l2jfree/branches/genesis@9305 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ l2jfree-core/src/main/java/com/l2jfree/gameserver/gameobjects/L2Item.java
@@ -41,7 +41,7 @@ public abstract class L2Item extends L2Object
 		return getTemplate().getName();
 	}
 	
-	public abstract int getCount();
+	public abstract long getCount();
 	
 	public abstract boolean isStackable();
